+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ### Changelog
 
 - [] .navbar-nav CSS adjusted to make collapsed navbar populate in a row and not column
+- [] Cards CSS adjusted on ABOUT page now are width 100 on mobile and width 75% on laptop/desktop screens
+- [] HomePage CSS adjusted for mobile and desktop.  Various spacing and sizing.
+- [] Kill Luke
+